Create honey-production.py
As you may have already heard, the honeybees are in a precarious state right now. You may have seen articles about the decline of the honeybee population for various reasons. in this project, I want to investigate this decline and how the trends of the past predict the future for the honeybees.